refactor(KERN): remove unused list components
