+++ app/components/content/ContentComponents.tsx
@@ -4,7 +4,7 @@ import KernSummaryOverviewSection from "~/components/kern/summaryOverview/Summar
 import { GridSection } from "~/components/layout/grid/GridSection";
 import type { StrapiContentComponent } from "~/services/cms/models/formElements/StrapiContentComponent";
 import { Grid } from "../layout/grid/Grid";
-import KernList from "../kern/KernList";
+import KernList from "./list/KernList";
 import KernHero from "../kern/KernHero";
 import KernTableOfContents from "../kern/KernTableOfContents";
 import KernBox from "../kern/KernBox";
